try again

+++ 0x00-MySQL_Advanced/7-average_score.sql
@@ -8,13 +8,17 @@ DELIMITER $$
 
 CREATE PROCEDURE ComputeAverageScoreForUser(IN user_id INT)
 BEGIN
-	DECLARE avg_score FLOAT;
+	DECLARE avg_score DECIMAL(5,2);
 
 	-- compute the average score
-	SELECT AVG(CAST(score AS DECIMAL)) INTO avg_score FROM corrections WHERE user_id = user_id;
+	SELECT AVG(CAST(score AS DECIMAL)) INTO avg_score
+	FROM corrections
+	WHERE user_id = user_id;
 
 	-- store the average score in the `users` table
-	UPDATE users SET average_score = avg_score WHERE id = user_id;
+	UPDATE users
+	SET average_score = avg_score
+	WHERE id = user_id;
 
 	-- Compute the average and store it into the `users` table directly
 	-- UPDATE users
